The CSS file has been moved to the page folder.
