+++ packages/client/src/router/index.tsx
@@ -6,7 +6,7 @@ import { PageLogin } from '../layout/pages/login/pageLogin'
 import { PageSignUp } from '../layout/pages/signUp/pageSignUp'
 import { PageProfile } from '../layout/pages/pageProfile'
 import { PageGame } from '../layout/pages/pageGame'
-import { PageLeaderboard } from '../layout/pages/pageLeaderboard'
+import { PageLeaderboard } from '../layout/pages/leaderboard/pageLeaderboard'
 import { PageForum } from '../layout/pages/pageForum'
 import { PageForumTopic } from '../layout/pages/pageForumTopic'
 import { Error500 } from '../layout/pages/page500/page500'
